Task7/requirements.txt
tensorflow-cpu
pandas
numpy==1.19.5
sklearn
flask
gunicorn
